pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = 'itd'
module-root = ''

[project]
name = "itd-sdk"
version = "2.0.0"
description = "ITD client for python"
readme = "README.md"
authors = [
  { name = "firedotguy", email = "nta16022013@gmail.com" }
]
license = "MIT"
dependencies = [
  "requests~=2.32.5",
  "pydantic~=2.12.5",
  "sseclient-py~=1.9.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
filetype = [
    "filetype>=1.2.0",
]
logs = [
    "rich>=15.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
]
